Convenient tbl_trees

Wrapper S7 class for the tbl_tree class of the tidytree package.

Usage

ConvenientTblTree

Format

An object of class S7_S3_class of length 3.

Details

Convenient tbl_tree objects are typically constructed with the makeTidyTree() function. In addition to usual tbl_tree objects, they also have a isTip column (boolean) and a y column that indicates the order in which the nodes would be plotted. This is needed by the orderWithTree() function that helps making heatmaps sorted in the same way.

Author

Charles Plessy