exDataFrame

Dataframe object containing species genome data.

Usage

data("exDataFrame")

Format

A data frame with 30 observations on the following 448 variables.

aligned_length_Min

a numeric vector

aligned_length_Q1

a numeric vector

aligned_length_Median

a numeric vector

aligned_length_Mean

a numeric vector

aligned_length_Q3

a numeric vector

aligned_length_Max

a numeric vector

aligned_length_L50

a numeric vector

aligned_length_Total

a numeric vector

aligned_length_N

a numeric vector

aligned_score_Min

a numeric vector

aligned_score_Q1

a numeric vector

aligned_score_Median

a numeric vector

aligned_score_Mean

a numeric vector

aligned_score_Q3

a numeric vector

aligned_score_Max

a numeric vector

aligned_score_L50

a numeric vector

aligned_score_Total

a numeric vector

aligned_score_N

a numeric vector

aligned_matches_Min

a numeric vector

aligned_matches_Q1

a numeric vector

aligned_matches_Median

a numeric vector

aligned_matches_Mean

a numeric vector

aligned_matches_Q3

a numeric vector

aligned_matches_Max

a numeric vector

aligned_matches_L50

a numeric vector

aligned_matches_Total

a numeric vector

aligned_matches_N

a numeric vector

aligned_mismatches_Min

a numeric vector

aligned_mismatches_Q1

a numeric vector

aligned_mismatches_Median

a numeric vector

aligned_mismatches_Mean

a numeric vector

aligned_mismatches_Q3

a numeric vector

aligned_mismatches_Max

a numeric vector

aligned_mismatches_L50

a numeric vector

aligned_mismatches_Total

a numeric vector

aligned_mismatches_N

a numeric vector

aligned_gaps_target_Q1

a numeric vector

aligned_gaps_target_Median

a numeric vector

aligned_gaps_target_Mean

a numeric vector

aligned_gaps_target_Q3

a numeric vector

aligned_gaps_target_Max

a numeric vector

aligned_gaps_target_L50

a numeric vector

aligned_gaps_target_Total

a numeric vector

aligned_gaps_target_N

a numeric vector

aligned_gaps_query_Q1

a numeric vector

aligned_gaps_query_Median

a numeric vector

aligned_gaps_query_Mean

a numeric vector

aligned_gaps_query_Q3

a numeric vector

aligned_gaps_query_Max

a numeric vector

aligned_gaps_query_L50

a numeric vector

aligned_gaps_query_Total

a numeric vector

aligned_gaps_query_N

a numeric vector

matching_aligned_Min

a numeric vector

matching_aligned_Q1

a numeric vector

matching_aligned_Median

a numeric vector

matching_aligned_Mean

a numeric vector

matching_aligned_Q3

a numeric vector

matching_aligned_Max

a numeric vector

matching_aligned_L50

a numeric vector

matching_aligned_Total

a numeric vector

matching_aligned_N

a numeric vector

matching_target_Min

a numeric vector

matching_target_Q1

a numeric vector

matching_target_Median

a numeric vector

matching_target_Mean

a numeric vector

matching_target_Q3

a numeric vector

matching_target_Max

a numeric vector

matching_target_L50

a numeric vector

matching_target_Total

a numeric vector

matching_target_N

a numeric vector

matching_query_Min

a numeric vector

matching_query_Q1

a numeric vector

matching_query_Median

a numeric vector

matching_query_Mean

a numeric vector

matching_query_Q3

a numeric vector

matching_query_Max

a numeric vector

matching_query_L50

a numeric vector

matching_query_Total

a numeric vector

matching_query_N

a numeric vector

mismatching_aligned_Min

a numeric vector

mismatching_aligned_Q1

a numeric vector

mismatching_aligned_Median

a numeric vector

mismatching_aligned_Mean

a numeric vector

mismatching_aligned_Q3

a numeric vector

mismatching_aligned_Max

a numeric vector

mismatching_aligned_L50

a numeric vector

mismatching_aligned_Total

a numeric vector

mismatching_aligned_N

a numeric vector

mismatching_target_Min

a numeric vector

mismatching_target_Q1

a numeric vector

mismatching_target_Median

a numeric vector

mismatching_target_Mean

a numeric vector

mismatching_target_Q3

a numeric vector

mismatching_target_Max

a numeric vector

mismatching_target_L50

a numeric vector

mismatching_target_Total

a numeric vector

mismatching_target_N

a numeric vector

mismatching_query_Min

a numeric vector

mismatching_query_Q1

a numeric vector

mismatching_query_Median

a numeric vector

mismatching_query_Mean

a numeric vector

mismatching_query_Q3

a numeric vector

mismatching_query_Max

a numeric vector

mismatching_query_L50

a numeric vector

mismatching_query_Total

a numeric vector

mismatching_query_N

a numeric vector

aligned_width_target_Min

a numeric vector

aligned_width_target_Q1

a numeric vector

aligned_width_target_Median

a numeric vector

aligned_width_target_Mean

a numeric vector

aligned_width_target_Q3

a numeric vector

aligned_width_target_Max

a numeric vector

aligned_width_target_L50

a numeric vector

aligned_width_target_Total

a numeric vector

aligned_width_target_N

a numeric vector

aligned_width_query_Min

a numeric vector

aligned_width_query_Q1

a numeric vector

aligned_width_query_Median

a numeric vector

aligned_width_query_Mean

a numeric vector

aligned_width_query_Q3

a numeric vector

aligned_width_query_Max

a numeric vector

aligned_width_query_L50

a numeric vector

aligned_width_query_Total

a numeric vector

aligned_width_query_N

a numeric vector

chain_width_target_Min

a numeric vector

chain_width_target_Q1

a numeric vector

chain_width_target_Median

a numeric vector

chain_width_target_Mean

a numeric vector

chain_width_target_Q3

a numeric vector

chain_width_target_Max

a numeric vector

chain_width_target_L50

a numeric vector

chain_width_target_Total

a numeric vector

chain_width_target_N

a numeric vector

chain_width_query_Min

a numeric vector

chain_width_query_Q1

a numeric vector

chain_width_query_Median

a numeric vector

chain_width_query_Mean

a numeric vector

chain_width_query_Q3

a numeric vector

chain_width_query_Max

a numeric vector

chain_width_query_L50

a numeric vector

chain_width_query_Total

a numeric vector

chain_width_query_N

a numeric vector

collinear_width_target_Min

a numeric vector

collinear_width_target_Q1

a numeric vector

collinear_width_target_Median

a numeric vector

collinear_width_target_Mean

a numeric vector

collinear_width_target_Q3

a numeric vector

collinear_width_target_Max

a numeric vector

collinear_width_target_L50

a numeric vector

collinear_width_target_Total

a numeric vector

collinear_width_target_N

a numeric vector

isolated_width_target_Min

a numeric vector

isolated_width_target_Q1

a numeric vector

isolated_width_target_Median

a numeric vector

isolated_width_target_Mean

a numeric vector

isolated_width_target_Q3

a numeric vector

isolated_width_target_Max

a numeric vector

isolated_width_target_L50

a numeric vector

isolated_width_target_Total

a numeric vector

isolated_width_target_N

a numeric vector

bridge_width_target_Min

a numeric vector

bridge_width_target_Q1

a numeric vector

bridge_width_target_Median

a numeric vector

bridge_width_target_Mean

a numeric vector

bridge_width_target_Q3

a numeric vector

bridge_width_target_Max

a numeric vector

bridge_width_target_L50

a numeric vector

bridge_width_target_Total

a numeric vector

bridge_width_target_N

a numeric vector

breakpoint_width_target_Min

a numeric vector

breakpoint_width_target_Q1

a numeric vector

breakpoint_width_target_Median

a numeric vector

breakpoint_width_target_Mean

a numeric vector

breakpoint_width_target_Q3

a numeric vector

breakpoint_width_target_Max

a numeric vector

breakpoint_width_target_L50

a numeric vector

breakpoint_width_target_Total

a numeric vector

breakpoint_width_target_N

a numeric vector

inverted_width_target_Min

a numeric vector

inverted_width_target_Q1

a numeric vector

inverted_width_target_Median

a numeric vector

inverted_width_target_Mean

a numeric vector

inverted_width_target_Q3

a numeric vector

inverted_width_target_Max

a numeric vector

inverted_width_target_L50

a numeric vector

inverted_width_target_Total

a numeric vector

inverted_width_target_N

a numeric vector

translocated_width_target_Min

a numeric vector

translocated_width_target_Q1

a numeric vector

translocated_width_target_Median

a numeric vector

translocated_width_target_Mean

a numeric vector

translocated_width_target_Q3

a numeric vector

translocated_width_target_Max

a numeric vector

translocated_width_target_L50

a numeric vector

translocated_width_target_Total

a numeric vector

translocated_width_target_N

a numeric vector

guessed_target_length_Min

a numeric vector

guessed_target_length_Q1

a numeric vector

guessed_target_length_Median

a numeric vector

guessed_target_length_Mean

a numeric vector

guessed_target_length_Q3

a numeric vector

guessed_target_length_Max

a numeric vector

guessed_target_length_L50

a numeric vector

guessed_target_length_Total

a numeric vector

guessed_target_length_N

a numeric vector

guessed_query_length_Min

a numeric vector

guessed_query_length_Q1

a numeric vector

guessed_query_length_Median

a numeric vector

guessed_query_length_Mean

a numeric vector

guessed_query_length_Q3

a numeric vector

guessed_query_length_Max

a numeric vector

guessed_query_length_L50

a numeric vector

guessed_query_length_Total

a numeric vector

guessed_query_length_N

a numeric vector

index_synteny_target

a numeric vector

index_synteny_query

a numeric vector

index_correlation_target

a numeric vector

index_correlation_query

a numeric vector

index_GOCvicinity4_target

a numeric vector

index_GOCvicinity4_query

a numeric vector

index_strandRand_target

a numeric vector

index_strandRand_query

a numeric vector

aligned_length_bestpair_Min

a numeric vector

aligned_length_bestpair_Q1

a numeric vector

aligned_length_bestpair_Median

a numeric vector

aligned_length_bestpair_Mean

a numeric vector

aligned_length_bestpair_Q3

a numeric vector

aligned_length_bestpair_Max

a numeric vector

aligned_length_bestpair_L50

a numeric vector

aligned_length_bestpair_Total

a numeric vector

aligned_length_bestpair_N

a numeric vector

aligned_score_bestpair_Min

a numeric vector

aligned_score_bestpair_Q1

a numeric vector

aligned_score_bestpair_Median

a numeric vector

aligned_score_bestpair_Mean

a numeric vector

aligned_score_bestpair_Q3

a numeric vector

aligned_score_bestpair_Max

a numeric vector

aligned_score_bestpair_L50

a numeric vector

aligned_score_bestpair_Total

a numeric vector

aligned_score_bestpair_N

a numeric vector

aligned_matches_bestpair_Min

a numeric vector

aligned_matches_bestpair_Q1

a numeric vector

aligned_matches_bestpair_Median

a numeric vector

aligned_matches_bestpair_Mean

a numeric vector

aligned_matches_bestpair_Q3

a numeric vector

aligned_matches_bestpair_Max

a numeric vector

aligned_matches_bestpair_L50

a numeric vector

aligned_matches_bestpair_Total

a numeric vector

aligned_matches_bestpair_N

a numeric vector

aligned_mismatches_bestpair_Min

a numeric vector

aligned_mismatches_bestpair_Q1

a numeric vector

aligned_mismatches_bestpair_Median

a numeric vector

aligned_mismatches_bestpair_Mean

a numeric vector

aligned_mismatches_bestpair_Q3

a numeric vector

aligned_mismatches_bestpair_Max

a numeric vector

aligned_mismatches_bestpair_L50

a numeric vector

aligned_mismatches_bestpair_Total

a numeric vector

aligned_mismatches_bestpair_N

a numeric vector

aligned_gaps_target_bestpair_Q1

a numeric vector

aligned_gaps_target_bestpair_Median

a numeric vector

aligned_gaps_target_bestpair_Mean

a numeric vector

aligned_gaps_target_bestpair_Q3

a numeric vector

aligned_gaps_target_bestpair_Max

a numeric vector

aligned_gaps_target_bestpair_L50

a numeric vector

aligned_gaps_target_bestpair_Total

a numeric vector

aligned_gaps_target_bestpair_N

a numeric vector

aligned_gaps_query_bestpair_Q1

a numeric vector

aligned_gaps_query_bestpair_Median

a numeric vector

aligned_gaps_query_bestpair_Mean

a numeric vector

aligned_gaps_query_bestpair_Q3

a numeric vector

aligned_gaps_query_bestpair_Max

a numeric vector

aligned_gaps_query_bestpair_L50

a numeric vector

aligned_gaps_query_bestpair_Total

a numeric vector

aligned_gaps_query_bestpair_N

a numeric vector

matching_aligned_bestpair_Min

a numeric vector

matching_aligned_bestpair_Q1

a numeric vector

matching_aligned_bestpair_Median

a numeric vector

matching_aligned_bestpair_Mean

a numeric vector

matching_aligned_bestpair_Q3

a numeric vector

matching_aligned_bestpair_Max

a numeric vector

matching_aligned_bestpair_L50

a numeric vector

matching_aligned_bestpair_Total

a numeric vector

matching_aligned_bestpair_N

a numeric vector

matching_target_bestpair_Min

a numeric vector

matching_target_bestpair_Q1

a numeric vector

matching_target_bestpair_Median

a numeric vector

matching_target_bestpair_Mean

a numeric vector

matching_target_bestpair_Q3

a numeric vector

matching_target_bestpair_Max

a numeric vector

matching_target_bestpair_L50

a numeric vector

matching_target_bestpair_Total

a numeric vector

matching_target_bestpair_N

a numeric vector

matching_query_bestpair_Min

a numeric vector

matching_query_bestpair_Q1

a numeric vector

matching_query_bestpair_Median

a numeric vector

matching_query_bestpair_Mean

a numeric vector

matching_query_bestpair_Q3

a numeric vector

matching_query_bestpair_Max

a numeric vector

matching_query_bestpair_L50

a numeric vector

matching_query_bestpair_Total

a numeric vector

matching_query_bestpair_N

a numeric vector

mismatching_aligned_bestpair_Min

a numeric vector

mismatching_aligned_bestpair_Q1

a numeric vector

mismatching_aligned_bestpair_Median

a numeric vector

mismatching_aligned_bestpair_Mean

a numeric vector

mismatching_aligned_bestpair_Q3

a numeric vector

mismatching_aligned_bestpair_Max

a numeric vector

mismatching_aligned_bestpair_L50

a numeric vector

mismatching_aligned_bestpair_Total

a numeric vector

mismatching_aligned_bestpair_N

a numeric vector

mismatching_target_bestpair_Min

a numeric vector

mismatching_target_bestpair_Q1

a numeric vector

mismatching_target_bestpair_Median

a numeric vector

mismatching_target_bestpair_Mean

a numeric vector

mismatching_target_bestpair_Q3

a numeric vector

mismatching_target_bestpair_Max

a numeric vector

mismatching_target_bestpair_L50

a numeric vector

mismatching_target_bestpair_Total

a numeric vector

mismatching_target_bestpair_N

a numeric vector

mismatching_query_bestpair_Min

a numeric vector

mismatching_query_bestpair_Q1

a numeric vector

mismatching_query_bestpair_Median

a numeric vector

mismatching_query_bestpair_Mean

a numeric vector

mismatching_query_bestpair_Q3

a numeric vector

mismatching_query_bestpair_Max

a numeric vector

mismatching_query_bestpair_L50

a numeric vector

mismatching_query_bestpair_Total

a numeric vector

mismatching_query_bestpair_N

a numeric vector

aligned_width_target_bestpair_Min

a numeric vector

aligned_width_target_bestpair_Q1

a numeric vector

aligned_width_target_bestpair_Median

a numeric vector

aligned_width_target_bestpair_Mean

a numeric vector

aligned_width_target_bestpair_Q3

a numeric vector

aligned_width_target_bestpair_Max

a numeric vector

aligned_width_target_bestpair_L50

a numeric vector

aligned_width_target_bestpair_Total

a numeric vector

aligned_width_target_bestpair_N

a numeric vector

aligned_width_query_bestpair_Min

a numeric vector

aligned_width_query_bestpair_Q1

a numeric vector

aligned_width_query_bestpair_Median

a numeric vector

aligned_width_query_bestpair_Mean

a numeric vector

aligned_width_query_bestpair_Q3

a numeric vector

aligned_width_query_bestpair_Max

a numeric vector

aligned_width_query_bestpair_L50

a numeric vector

aligned_width_query_bestpair_Total

a numeric vector

aligned_width_query_bestpair_N

a numeric vector

chain_width_target_bestpair_Min

a numeric vector

chain_width_target_bestpair_Q1

a numeric vector

chain_width_target_bestpair_Median

a numeric vector

chain_width_target_bestpair_Mean

a numeric vector

chain_width_target_bestpair_Q3

a numeric vector

chain_width_target_bestpair_Max

a numeric vector

chain_width_target_bestpair_L50

a numeric vector

chain_width_target_bestpair_Total

a numeric vector

chain_width_target_bestpair_N

a numeric vector

chain_width_query_bestpair_Min

a numeric vector

chain_width_query_bestpair_Q1

a numeric vector

chain_width_query_bestpair_Median

a numeric vector

chain_width_query_bestpair_Mean

a numeric vector

chain_width_query_bestpair_Q3

a numeric vector

chain_width_query_bestpair_Max

a numeric vector

chain_width_query_bestpair_L50

a numeric vector

chain_width_query_bestpair_Total

a numeric vector

chain_width_query_bestpair_N

a numeric vector

collinear_width_target_bestpair_Min

a numeric vector

collinear_width_target_bestpair_Q1

a numeric vector

collinear_width_target_bestpair_Median

a numeric vector

collinear_width_target_bestpair_Mean

a numeric vector

collinear_width_target_bestpair_Q3

a numeric vector

collinear_width_target_bestpair_Max

a numeric vector

collinear_width_target_bestpair_L50

a numeric vector

collinear_width_target_bestpair_Total

a numeric vector

collinear_width_target_bestpair_N

a numeric vector

isolated_width_target_bestpair_Min

a numeric vector

isolated_width_target_bestpair_Q1

a numeric vector

isolated_width_target_bestpair_Median

a numeric vector

isolated_width_target_bestpair_Mean

a numeric vector

isolated_width_target_bestpair_Q3

a numeric vector

isolated_width_target_bestpair_Max

a numeric vector

isolated_width_target_bestpair_L50

a numeric vector

isolated_width_target_bestpair_Total

a numeric vector

isolated_width_target_bestpair_N

a numeric vector

bridge_width_target_bestpair_Min

a numeric vector

bridge_width_target_bestpair_Q1

a numeric vector

bridge_width_target_bestpair_Median

a numeric vector

bridge_width_target_bestpair_Mean

a numeric vector

bridge_width_target_bestpair_Q3

a numeric vector

bridge_width_target_bestpair_Max

a numeric vector

bridge_width_target_bestpair_L50

a numeric vector

bridge_width_target_bestpair_Total

a numeric vector

bridge_width_target_bestpair_N

a numeric vector

breakpoint_width_target_bestpair_Min

a numeric vector

breakpoint_width_target_bestpair_Q1

a numeric vector

breakpoint_width_target_bestpair_Median

a numeric vector

breakpoint_width_target_bestpair_Mean

a numeric vector

breakpoint_width_target_bestpair_Q3

a numeric vector

breakpoint_width_target_bestpair_Max

a numeric vector

breakpoint_width_target_bestpair_L50

a numeric vector

breakpoint_width_target_bestpair_Total

a numeric vector

breakpoint_width_target_bestpair_N

a numeric vector

inverted_width_target_bestpair_Min

a numeric vector

inverted_width_target_bestpair_Q1

a numeric vector

inverted_width_target_bestpair_Median

a numeric vector

inverted_width_target_bestpair_Mean

a numeric vector

inverted_width_target_bestpair_Q3

a numeric vector

inverted_width_target_bestpair_Max

a numeric vector

inverted_width_target_bestpair_L50

a numeric vector

inverted_width_target_bestpair_Total

a numeric vector

inverted_width_target_bestpair_N

a numeric vector

translocated_width_target_bestpair_Min

a numeric vector

translocated_width_target_bestpair_Q1

a numeric vector

translocated_width_target_bestpair_Median

a numeric vector

translocated_width_target_bestpair_Mean

a numeric vector

translocated_width_target_bestpair_Q3

a numeric vector

translocated_width_target_bestpair_Max

a numeric vector

translocated_width_target_bestpair_L50

a numeric vector

translocated_width_target_bestpair_Total

a numeric vector

translocated_width_target_bestpair_N

a numeric vector

guessed_target_length_bestpair_Min

a numeric vector

guessed_target_length_bestpair_Q1

a numeric vector

guessed_target_length_bestpair_Median

a numeric vector

guessed_target_length_bestpair_Mean

a numeric vector

guessed_target_length_bestpair_Q3

a numeric vector

guessed_target_length_bestpair_Max

a numeric vector

guessed_target_length_bestpair_L50

a numeric vector

guessed_target_length_bestpair_Total

a numeric vector

guessed_target_length_bestpair_N

a numeric vector

guessed_query_length_bestpair_Min

a numeric vector

guessed_query_length_bestpair_Q1

a numeric vector

guessed_query_length_bestpair_Median

a numeric vector

guessed_query_length_bestpair_Mean

a numeric vector

guessed_query_length_bestpair_Q3

a numeric vector

guessed_query_length_bestpair_Max

a numeric vector

guessed_query_length_bestpair_L50

a numeric vector

guessed_query_length_bestpair_Total

a numeric vector

guessed_query_length_bestpair_N

a numeric vector

index_synteny_target_bestpair

a numeric vector

index_synteny_query_bestpair

a numeric vector

index_correlation_target_bestpair

a numeric vector

index_correlation_query_bestpair

a numeric vector

index_GOCvicinity4_target_bestpair

a numeric vector

index_GOCvicinity4_query_bestpair

a numeric vector

index_strandRand_target_bestpair

a numeric vector

index_strandRand_query_bestpair

a numeric vector

species1

a character vector

species2

a character vector

index_avg_strandRand

a numeric vector

percent_identity_global

a numeric vector

Details

Dataframe containing a wide variety of genome comparison calculations. Calculations ending with "..._bestpair" are meant to be utilized for bacteria, archaea, and other species with many plasmid, minichromosomes, etc. The bestpair calculation compares the two largest chromosomes of each species rather than the entire genome.

Source

Data comes from Nextflow pairwise alignment pipeline.

Examples

data(exDataFrame)
## maybe str(exDataFrame) ; plot(exDataFrame) ...