Format species data and calculate percent identity and strand randomisation variables as part of the dataframe.

Format species data and calculate percent identity and strand randomisation variables as part of the dataframe.

Usage

formatStats(files)

Arguments

files

A named character vector of paths to YAML files. The names will become the row names of the output data frame. Compressed files are\ accepted.

Value

Formatted dataframe with strand randomisation and percent identities between species calculated.

Examples

resultsDir <- system.file("extdata/PairwiseComparisons", package = "ScrambledTreeBuilder")
yamlFileData <- list.files(resultsDir, pattern = "*.yaml.bz2", full.names = TRUE)
names(yamlFileData) <- yamlFileData |> basename() |> sub(pat = ".yaml.bz2", rep="")
exDataFrame <- formatStats(yamlFileData)
exDataFrame[1:10,1:6]
#>                                                             aligned_length_Min
#> Halobacterium_litoreum___Halobacterium_noricense                            62
#> Halobacterium_litoreum___Halobacterium_salinarum                           100
#> Halobacterium_litoreum___Haloferax_mediterranei_ATCC_33500                  92
#> Halobacterium_litoreum___Haloferax_volcanii_DS2                             72
#> Halobacterium_litoreum___Salarchaeum_japonicum                              66
#> Halobacterium_noricense___Halobacterium_litoreum                            66
#> Halobacterium_noricense___Halobacterium_salinarum                           76
#> Halobacterium_noricense___Haloferax_mediterranei_ATCC_33500                 60
#> Halobacterium_noricense___Haloferax_volcanii_DS2                            79
#> Halobacterium_noricense___Salarchaeum_japonicum                             77
#>                                                             aligned_length_Q1
#> Halobacterium_litoreum___Halobacterium_noricense                       717.50
#> Halobacterium_litoreum___Halobacterium_salinarum                       778.00
#> Halobacterium_litoreum___Haloferax_mediterranei_ATCC_33500             546.50
#> Halobacterium_litoreum___Haloferax_volcanii_DS2                        585.00
#> Halobacterium_litoreum___Salarchaeum_japonicum                         674.75
#> Halobacterium_noricense___Halobacterium_litoreum                       718.75
#> Halobacterium_noricense___Halobacterium_salinarum                      731.00
#> Halobacterium_noricense___Haloferax_mediterranei_ATCC_33500            557.00
#> Halobacterium_noricense___Haloferax_volcanii_DS2                       577.75
#> Halobacterium_noricense___Salarchaeum_japonicum                        684.00
#>                                                             aligned_length_Median
#> Halobacterium_litoreum___Halobacterium_noricense                           1409.0
#> Halobacterium_litoreum___Halobacterium_salinarum                           1459.5
#> Halobacterium_litoreum___Haloferax_mediterranei_ATCC_33500                  949.0
#> Halobacterium_litoreum___Haloferax_volcanii_DS2                             994.5
#> Halobacterium_litoreum___Salarchaeum_japonicum                             1260.5
#> Halobacterium_noricense___Halobacterium_litoreum                           1422.5
#> Halobacterium_noricense___Halobacterium_salinarum                          1349.0
#> Halobacterium_noricense___Haloferax_mediterranei_ATCC_33500                 963.0
#> Halobacterium_noricense___Haloferax_volcanii_DS2                            998.0
#> Halobacterium_noricense___Salarchaeum_japonicum                            1235.0
#>                                                             aligned_length_Mean
#> Halobacterium_litoreum___Halobacterium_noricense                       2177.722
#> Halobacterium_litoreum___Halobacterium_salinarum                       2248.691
#> Halobacterium_litoreum___Haloferax_mediterranei_ATCC_33500             1185.227
#> Halobacterium_litoreum___Haloferax_volcanii_DS2                        1296.762
#> Halobacterium_litoreum___Salarchaeum_japonicum                         1813.424
#> Halobacterium_noricense___Halobacterium_litoreum                       2211.831
#> Halobacterium_noricense___Halobacterium_salinarum                      2089.673
#> Halobacterium_noricense___Haloferax_mediterranei_ATCC_33500            1213.886
#> Halobacterium_noricense___Haloferax_volcanii_DS2                       1309.893
#> Halobacterium_noricense___Salarchaeum_japonicum                        1781.424
#>                                                             aligned_length_Q3
#> Halobacterium_litoreum___Halobacterium_noricense                      2926.50
#> Halobacterium_litoreum___Halobacterium_salinarum                      3053.00
#> Halobacterium_litoreum___Haloferax_mediterranei_ATCC_33500            1506.50
#> Halobacterium_litoreum___Haloferax_volcanii_DS2                       1649.25
#> Halobacterium_litoreum___Salarchaeum_japonicum                        2315.00
#> Halobacterium_noricense___Halobacterium_litoreum                      2988.50
#> Halobacterium_noricense___Halobacterium_salinarum                     2668.50
#> Halobacterium_noricense___Haloferax_mediterranei_ATCC_33500           1521.00
#> Halobacterium_noricense___Haloferax_volcanii_DS2                      1640.25
#> Halobacterium_noricense___Salarchaeum_japonicum                       2238.50
#>                                                             aligned_length_Max
#> Halobacterium_litoreum___Halobacterium_noricense                         18125
#> Halobacterium_litoreum___Halobacterium_salinarum                         18237
#> Halobacterium_litoreum___Haloferax_mediterranei_ATCC_33500               14391
#> Halobacterium_litoreum___Haloferax_volcanii_DS2                          14440
#> Halobacterium_litoreum___Salarchaeum_japonicum                           20211
#> Halobacterium_noricense___Halobacterium_litoreum                         18125
#> Halobacterium_noricense___Halobacterium_salinarum                        20473
#> Halobacterium_noricense___Haloferax_mediterranei_ATCC_33500              14665
#> Halobacterium_noricense___Haloferax_volcanii_DS2                         13872
#> Halobacterium_noricense___Salarchaeum_japonicum                          18158