Pairwise comparison matrix

Make a matrix from a data frame containing species data.

Usage

makeMatrix(DF, column, defaultDiagonal = 100, defaultValue = NA)

Arguments

DF

Data frame containing species information.

column

The variable from the data frame you want to extract species data from.

defaultDiagonal

Maximum possible value of variable.

defaultValue

Minimum possible value of variable.

Value

A square matrix where rows and columns are sample identifiers and values are statistics for the pairs of samples. In addition it has an attribute "builtWith" that records the value of column.

Author

Noa Brenner

Charles Plessy

Examples

makeMatrix(Halo_DF, "percent_identity_global", 100, 50)
#>                          species2
#> species1                  Halobacterium_litoreum Halobacterium_noricense
#>   Halobacterium_litoreum               100.00000                78.81199
#>   Halobacterium_noricense               78.73936               100.00000
#>   Halobacterium_salinarum               75.84298                74.94759
#>   Haloferax_mediterranei                68.27405                66.42841
#>   Haloferax_volcanii                    70.11358                69.93183
#>   Salarchaeum_japonicum                 72.75352                72.11068
#>                          species2
#> species1                  Halobacterium_salinarum Haloferax_mediterranei
#>   Halobacterium_litoreum                 75.75208               68.72077
#>   Halobacterium_noricense                74.94623               68.45333
#>   Halobacterium_salinarum               100.00000               67.00552
#>   Haloferax_mediterranei                 66.97889              100.00000
#>   Haloferax_volcanii                     68.48311               80.39441
#>   Salarchaeum_japonicum                  71.80855               68.65929
#>                          species2
#> species1                  Haloferax_volcanii Salarchaeum_japonicum
#>   Halobacterium_litoreum            70.16293              72.74950
#>   Halobacterium_noricense           69.96147              72.42201
#>   Halobacterium_salinarum           68.48878              71.55815
#>   Haloferax_mediterranei            80.22559              68.41933
#>   Haloferax_volcanii               100.00000              69.79642
#>   Salarchaeum_japonicum             69.82446             100.00000