Make new tibble dataset with added value column from a matrix

Make new tibble dataset with added value column from a matrix

Usage

makeValueTibble(your_tibble, your_matrix, fun = mean, colname = "value")

Arguments

your_tibble

Original tibble dataset to add value column to.

your_matrix

Matrix containing values to add to tibble.

Value

New tibble with values added.

Examples

makeValueTibble(Tibble, valueMatrix)
#> # A tbl_tree abstraction: 11 × 5
#> # which can be converted to treedata or phylo 
#> # via as.treedata or as.phylo
#>    parent  node branch.length label                               value
#>     <int> <int>         <dbl> <chr>                               <dbl>
#>  1     11     1         15.0  Halobacterium_noricense           NA     
#>  2     10     2         18.0  Halobacterium_salinarum           NA     
#>  3      8     3         13.7  Haloferax_mediterranei_ATCC_33500 NA     
#>  4      8     4         13.7  Haloferax_volcanii_DS2            NA     
#>  5      9     5         20.1  Salarchaeum_japonicum             NA     
#>  6     11     6         15.0  Halobacterium_litoreum            NA     
#>  7      7     7         NA    NA                                 0.0666
#>  8      7     8         11.1  NA                                 0.337 
#>  9      7     9          4.62 NA                                 0.121 
#> 10      9    10          2.19 NA                                 0.387 
#> 11     10    11          2.98 NA                                 0.397