Oikopleura data

Oikopleura data points to overlay on nucleotide distance / scrambling index plots.

Usage

oikData

Format

List of tibbles. The name of each tibble is just a time stamp.

Details

This test dataset is derived from genome alignments produced by a beta version of the https://nf-co.re/pairgenomealign pipeline.

2025_02_17 6 distances using the RY128 seed.

2025_02_25 10 distances using the RY4 seed.

See also

Author

Charles Plessy

Examples

oikData
#> $`2025_02_17`
#> # A tibble: 6 × 7
#>    MRCA     x      y    xerr      yerr     n clade     
#>   <dbl> <dbl>  <dbl>   <dbl>     <dbl> <int> <chr>     
#> 1     8  78.8 0.0905 0.252   0.0183       20 Oikopleura
#> 2     9  96.7 0.854  0.00424 0.000352      2 Oikopleura
#> 3    10  87.5 0.242  0.0572  0.0370        8 Oikopleura
#> 4    11  97.1 0.478  0.0628  0.148         6 Oikopleura
#> 5    12  97.4 0.273  0.0624  0.0804        4 Oikopleura
#> 6    13  97.7 0.656  0.00233 0.0000759     2 Oikopleura
#> 
#> $`2025_02_25`
#> # A tibble: 10 × 7
#>     MRCA     x      y    xerr     yerr     n clade     
#>    <dbl> <dbl>  <dbl>   <dbl>    <dbl> <int> <chr>     
#>  1    12  78.2 0.0744 0.197   0.0186      36 Oikopleura
#>  2    13  96.6 0.848  0.00304 0.000250     2 Oikopleura
#>  3    14  86.9 0.199  0.176   0.0412      36 Oikopleura
#>  4    15  97.4 0.665  0.0197  0.0563       4 Oikopleura
#>  5    16  97.1 0.510  0.105   0.119       10 Oikopleura
#>  6    17  97.4 0.269  0.0660  0.0531       8 Oikopleura
#>  7    18  97.6 0.625  0.0242  0.0320       6 Oikopleura
#>  8    19  98.8 0.633  0.00283 0.000416     2 Oikopleura
#>  9    20  97.7 0.631  0.0260  0.0278       4 Oikopleura
#> 10    21  97.9 0.684  0.00382 0.000751     2 Oikopleura
#>