Subset a matrix with a subtree

Given a matrix of values where the rows and column names are species names that are labels in a tree, subset that matrix to the species that are present in the subtree starting at a given node.

Usage

subMatrix(tree_tibble, m, node = NA, simplify.names = FALSE)

Arguments

tree_tibble

A tree in the tbl_tree class of the tidytree package.

m

A matrix, typically the output of makeMatrix().

node

The ID number of the node to base the subtree.

simplify.names

Reduce the genus name to a single letter.

Value

A subset of the m matrix.

Examples

subMatrix(tibbleWithValue, valueMatrix, simpl = TRUE)
#>                            H_noricense  H_salinarum H_mediterranei_ATCC_33500
#> H_noricense               100.00000000   0.23860920                0.04765400
#> H_salinarum                 0.23871495 100.00000000                0.07367205
#> H_mediterranei_ATCC_33500   0.04880815   0.07503800              100.00000000
#> H_volcanii_DS2              0.02784920   0.13530080                0.33807060
#> S_japonicum                 0.11476675   0.06630175                0.10473430
#> H_litoreum                  0.39885460   0.53733160                0.07850880
#>                           H_volcanii_DS2 S_japonicum   H_litoreum
#> H_noricense                   0.02749270   0.1146204   0.39714395
#> H_salinarum                   0.13299345   0.0671487   0.53626875
#> H_mediterranei_ATCC_33500     0.33675900   0.1063006   0.08857840
#> H_volcanii_DS2              100.00000000   0.0044876   0.04606195
#> S_japonicum                   0.00848495 100.0000000   0.18109200
#> H_litoreum                    0.04474345   0.1884153 100.00000000
subMatrix(tibbleWithValue, valueMatrix, 9)
#>                         Halobacterium_noricense Halobacterium_salinarum
#> Halobacterium_noricense             100.0000000              0.23860920
#> Halobacterium_salinarum               0.2387149            100.00000000
#> Salarchaeum_japonicum                 0.1147668              0.06630175
#> Halobacterium_litoreum                0.3988546              0.53733160
#>                         Salarchaeum_japonicum Halobacterium_litoreum
#> Halobacterium_noricense             0.1146204              0.3971440
#> Halobacterium_salinarum             0.0671487              0.5362688
#> Salarchaeum_japonicum             100.0000000              0.1810920
#> Halobacterium_litoreum              0.1884153            100.0000000
subTree(tibbleWithValue, 9) |> subMatrix(valueMatrix)
#>                         Halobacterium_noricense Halobacterium_salinarum
#> Halobacterium_noricense             100.0000000              0.23860920
#> Halobacterium_salinarum               0.2387149            100.00000000
#> Salarchaeum_japonicum                 0.1147668              0.06630175
#> Halobacterium_litoreum                0.3988546              0.53733160
#>                         Salarchaeum_japonicum Halobacterium_litoreum
#> Halobacterium_noricense             0.1146204              0.3971440
#> Halobacterium_salinarum             0.0671487              0.5362688
#> Salarchaeum_japonicum             100.0000000              0.1810920
#> Halobacterium_litoreum              0.1884153            100.0000000