Make a matrix from a data frame containing species data.

Make a matrix from a data frame containing species data.

Usage

makeMatrix(DF, column, defaultDiagonal = 100, defaultValue = NA)

Arguments

DF

Data frame containing species information.

column

The variable from the data frame you want to extract species data from.

defaultDiagonal

Maximum possible value of variable.

defaultValue

Minimum possible value of variable.

Value

A matrix containing species names and data for the selected variable.

Examples

makeMatrix(exDataFrame, "percent_identity_global", 100, 50)
#>                                   Halobacterium_noricense
#> Halobacterium_noricense                         100.00000
#> Halobacterium_salinarum                          74.74168
#> Haloferax_mediterranei_ATCC_33500                66.51318
#> Haloferax_volcanii_DS2                           69.96063
#> Salarchaeum_japonicum                            71.94129
#> Halobacterium_litoreum                           78.87231
#>                                   Halobacterium_salinarum
#> Halobacterium_noricense                          74.72164
#> Halobacterium_salinarum                         100.00000
#> Haloferax_mediterranei_ATCC_33500                67.06152
#> Haloferax_volcanii_DS2                           68.47755
#> Salarchaeum_japonicum                            71.82141
#> Halobacterium_litoreum                           75.78684
#>                                   Haloferax_mediterranei_ATCC_33500
#> Halobacterium_noricense                                    68.55420
#> Halobacterium_salinarum                                    67.06788
#> Haloferax_mediterranei_ATCC_33500                         100.00000
#> Haloferax_volcanii_DS2                                     80.98460
#> Salarchaeum_japonicum                                      68.70265
#> Halobacterium_litoreum                                     68.88111
#>                                   Haloferax_volcanii_DS2 Salarchaeum_japonicum
#> Halobacterium_noricense                         69.98252              72.22600
#> Halobacterium_salinarum                         68.48654              71.57797
#> Haloferax_mediterranei_ATCC_33500               80.80996              68.50482
#> Haloferax_volcanii_DS2                         100.00000              69.84616
#> Salarchaeum_japonicum                           69.87968             100.00000
#> Halobacterium_litoreum                          70.25591              72.76736
#>                                   Halobacterium_litoreum
#> Halobacterium_noricense                         78.80471
#> Halobacterium_salinarum                         75.89491
#> Haloferax_mediterranei_ATCC_33500               68.49688
#> Haloferax_volcanii_DS2                          70.22052
#> Salarchaeum_japonicum                           72.77139
#> Halobacterium_litoreum                         100.00000