Make a matrix from a data frame containing species data.

Make a matrix from a data frame containing species data.

Usage

makeMatrix(DF, column, defaultDiagonal = 100, defaultValue = NA)

Arguments

DF

Data frame containing species information.

column

The variable from the data frame you want to extract species data from.

defaultDiagonal

Maximum possible value of variable.

defaultValue

Minimum possible value of variable.

Value

A matrix containing species names and data for the selected variable.

Examples

makeMatrix(Halo_DF, "percent_identity_global", 100, 50)
#>                         Halobacterium_noricense Halobacterium_salinarum
#> Halobacterium_noricense               100.00000                74.94623
#> Halobacterium_salinarum                74.94759               100.00000
#> Haloferax_mediterranei                 66.42841                66.97889
#> Haloferax_volcanii                     69.93183                68.48311
#> Salarchaeum_japonicum                  72.11068                71.80855
#> Halobacterium_litoreum                 78.81199                75.75208
#>                         Haloferax_mediterranei Haloferax_volcanii
#> Halobacterium_noricense               68.45333           69.96147
#> Halobacterium_salinarum               67.00552           68.48878
#> Haloferax_mediterranei               100.00000           80.22559
#> Haloferax_volcanii                    80.39441          100.00000
#> Salarchaeum_japonicum                 68.65929           69.82446
#> Halobacterium_litoreum                68.72077           70.16293
#>                         Salarchaeum_japonicum Halobacterium_litoreum
#> Halobacterium_noricense              72.42201               78.73936
#> Halobacterium_salinarum              71.55815               75.84298
#> Haloferax_mediterranei               68.41933               68.27405
#> Haloferax_volcanii                   69.79642               70.11358
#> Salarchaeum_japonicum               100.00000               72.75352
#> Halobacterium_litoreum               72.74950              100.00000