Subset a matrix with a subtree

Given a matrix of values where the rows and column names are species names that are labels in a tree, subset that matrix to the species that are present in the subtree starting at a given node.

Usage

subMatrix(tree_tibble, m, node = NA, simplify.names = FALSE)

Arguments

tree_tibble

A tree in the tbl_tree class of the tidytree package.

m

A matrix, typically the output of makeMatrix().

node

The ID number of the node to base the subtree.

simplify.names

Reduce the genus name to a single letter.

Value

A subset of the m matrix.

Examples

subMatrix(Halo_Tree, Halo_PercentDiff, simpl = TRUE)
#>                H_noricense H_salinarum H_mediterranei H_volcanii S_japonicum
#> H_noricense        0.00000    25.05377       31.54667   30.03853    27.57799
#> H_salinarum       25.05241     0.00000       32.99448   31.51122    28.44185
#> H_mediterranei    33.57159    33.02111        0.00000   19.77441    31.58067
#> H_volcanii        30.06817    31.51689       19.60559    0.00000    30.20358
#> S_japonicum       27.88932    28.19145       31.34071   30.17554     0.00000
#> H_litoreum        21.18801    24.24792       31.27923   29.83707    27.25050
#>                H_litoreum
#> H_noricense      21.26064
#> H_salinarum      24.15702
#> H_mediterranei   31.72595
#> H_volcanii       29.88642
#> S_japonicum      27.24648
#> H_litoreum        0.00000
subMatrix(Halo_Tree, Halo_PercentDiff, 9)
#>                         Halobacterium_noricense Halobacterium_salinarum
#> Halobacterium_noricense                 0.00000                25.05377
#> Halobacterium_salinarum                25.05241                 0.00000
#> Salarchaeum_japonicum                  27.88932                28.19145
#> Halobacterium_litoreum                 21.18801                24.24792
#>                         Salarchaeum_japonicum Halobacterium_litoreum
#> Halobacterium_noricense              27.57799               21.26064
#> Halobacterium_salinarum              28.44185               24.15702
#> Salarchaeum_japonicum                 0.00000               27.24648
#> Halobacterium_litoreum               27.25050                0.00000
subTree(Halo_Tree, 9) |> subMatrix(Halo_PercentDiff)
#>                         Halobacterium_noricense Halobacterium_salinarum
#> Halobacterium_noricense                 0.00000                25.05377
#> Halobacterium_salinarum                25.05241                 0.00000
#> Salarchaeum_japonicum                  27.88932                28.19145
#> Halobacterium_litoreum                 21.18801                24.24792
#>                         Salarchaeum_japonicum Halobacterium_litoreum
#> Halobacterium_noricense              27.57799               21.26064
#> Halobacterium_salinarum              28.44185               24.15702
#> Salarchaeum_japonicum                 0.00000               27.24648
#> Halobacterium_litoreum               27.25050                0.00000